move disabled tests to one block and moved eventscope tests
